+++ src/components/Searchbar/SearchBar.jsx
@@ -1,34 +1,14 @@
-import { useState } from "react";
+import { useSearch } from "../../contexts/SearchCityContext";
 
 export default function SearchBar() {
-  const [searchCity, setSearchCity] = useState("");
-  const [cityDataSearched, setcityDataSearched] = useState({});
-
-  const handlerOnChange = (e) => setSearchCity(e.target.value);
-
-  const handleOnSubmit = (e) => {
-    e.preventDefault(); 
-    fetch(
-      `https://api.openweathermap.org/data/2.5/weather?q=${searchCity}&units=metric&lang=fr&appid=${
-        import.meta.env.VITE_APP_ID
-      }`
-    )
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error("Failed to fetch current data");
-        }
-        return response.json();
-      })
-      .then((cityData) => {
-        setcityDataSearched(cityData);
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
+  const { searchCity, cityDataSearched, handleOnSubmit, handlerOnChange } =
+    useSearch();
 
   console.info("search: " + searchCity);
-  console.info("fetched city: " + cityDataSearched.main.temp);
+  console.info(
+    "fetched city: " + cityDataSearched.name,
+    cityDataSearched.main.temp
+  );
 
   return (
     <form onSubmit={handleOnSubmit}>
