+++ src/main.jsx
@@ -1,10 +1,14 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import App from './App.jsx'
-import './styles/index.scss'
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { SearchProvider } from "./contexts/SearchCityContext.jsx";
 
-createRoot(document.getElementById('root')).render(
+import App from "./App.jsx";
+import "./styles/index.scss";
+
+createRoot(document.getElementById("root")).render(
   <StrictMode>
-    <App />
-  </StrictMode>,
-)
+    <SearchProvider>
+      <App />
+    </SearchProvider>
+  </StrictMode>
+);
